增加权限

+++ app/Http/Controllers/UserAddressesController.php
@@ -23,6 +23,7 @@ class UserAddressesController extends Controller
 
     public function edit(UserAddress $user_address)
     {   
+        $this->authorize('own', $user_address);
         return view('user_addresses.create_and_edit', ['address' => $user_address]);
     }
 
@@ -42,7 +43,8 @@ class UserAddressesController extends Controller
     }
 
     public function update(UserAddress $user_address, UserAddressRequest $request)
-    {
+    {   
+        $this->authorize('own', $user_address);
         $user_address->update($request->only([
             'province',
             'city',
@@ -57,10 +59,10 @@ class UserAddressesController extends Controller
     }
 
     public function destroy(UserAddress $user_address)
-    {
+    {   
+        $this->authorize('own', $user_address);
         $user_address->delete();
-
-         return [];
+        return [];
     }
 
 }
